+++ src/main/java/com/telepathicgrunt/repurposedstructures/configs/RSMineshaftsConfig.java
@@ -75,9 +75,9 @@ public class RSMineshaftsConfig {
 
         builder.comment("-----------------------------------------------------------------------------------------",
                 " Controls the probability of spawning a Repurposed Structures Mineshafts per chunk.",
-                " Will replace Vanilla Mineshafts in the biomes with Repurposed Structures's themed Mineshafts.",
+                " Will replace Vanilla Mineshafts with RS's themed Mineshafts if both would've been in same biome.",
                 " 0 is no Mineshafts while 1000 is max spawnrate.",
-                " Note: Vanilla Mineshafts will start spawning when this is set to 0 and game is restarted.");
+                " Note: Vanilla Mineshafts will spawn again if a RS Mineshafts's entry is set to 0 for the biome.");
         builder.push("Spawnrate");
 
         birchMineshaftSpawnrate = builder